feddit-anaiyzer/pyproject.toml
[tool.poetry]
name = "feddit_analyzer"
version = "0.1.0"
description = "Web API that identifies if comments on a given subfeddit or category are positive or negative."
authors = ["Daniel, Martín Martínez <danitiana98@gmail.com>"]
packages = [{include = "feddit_analyzer"}]

license = "MIT"

keywords = [""]

repository = "https://github.com/dainelli98/feddit-analyzer.git"

readme = ["README.md", "RELEASE_NOTES.md"]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English"
]

include = ["feddit_analyzer/*"]
exclude = ["tests/*"]

[tool.poetry.dependencies]
python = "~3.11"
cachetools = "^5.3.3"
click = ">=8"
fastapi = "^0.111.0"
loguru = ">=0.6,<1.0"
pydantic = "^2.7.3"
python-dotenv = "^1.0.1"
typing-extensions = "^4.12.2"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "24.4.2"}
coverage = ">=6.4.3"
cruft = ">=2.12.0"
docformatter = "1.7.5"
GitPython = ">=3.1.27"
jupyterlab = ">=3.2.5"
jupytext = ">=1.13.7"
jupyter_contrib_nbextensions = ">=0.5.1"
myst-parser = ">=0.18.1"
nbsphinx = "^0.9.2"
pre-commit = ">=2.20.0"
pydata-sphinx-theme = ">=0.14.0"
pytest = ">=7.2.0"
pytest-asyncio = "^0.23.7"
pytest-cov = ">=3.0.0"
pytest-httpx = "^0.30.0"
pytest-mock = ">=1.7.1,<4.0"
pytest-runner = ">=6.0.0"
python-dotenv = ">=0.20.0"
pyupgrade = "3.15.2"
requests-mock = ">=1.0.0"
ruff = "0.4.8"
sphinx = ">=5.3.0"
sphinx-autodoc-typehints = ">=1.21.1"
sphinx-copybutton = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
feddit-analyzer = "feddit_analyzer.__main__:_main"
